backend/requirements.txt
Django==4.2.20
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
channels==4.0.0